refactor(forge): split Forgejo auth methods into forgejo_auth.go (#1025)
